Check contracts for payment acceptances
If a contract has no transitions which accept payment, raise a warning.

Also if a transition has any code path with potentially multiple
"accept" statements in it, raise a warning.  We do not raise an error
since it is not possible via static analysis to know for sure in all
cases whether multiple "accept" statements would be reached if
present, since this can be dependent on conditions which are only
known at run-time.  As a silly example, this transition may or may not
attempt to accept twice:

    transition donate_once_or_twice(twice : Bool)
      accept;
      match twice with
      | True => accept
      | False =>
        e = { _eventname : "Thanks" };
        event e
      end
    end

In contrast, this transition is guaranteed to accept only once:

    transition donate_once(switch : Bool)
      match switch with
      | True => accept
      | False =>
        e = { _eventname : "Thanks" };
        event e
      end;
      match switch with
      | False => accept
      | True =>
        e = { _eventname : "Thanks" };
        event e
      end
    end

however it would be very difficult to implement analysis which catches
cases like this.